requirements.txt
streamlit==1.38.0
python-dotenv==1.0.1

SQLAlchemy==2.0.34
pymysql==1.1.1

langgraph==0.2.46

openai==1.51.2
httpx==0.27.2

pinecone-client==5.0.1

pydantic==2.8.2
tqdm==4.66.5

cryptography>=41.0.0
